split utility vector handlers into utils module
